Handle if database does not exist, add readme to both parts of project.
The reason for having two readme's is so that one is viewable on load of
the repository, just to give a glimpse into the project. I understand
that it should accoridng to requirements go into the CodingTracker/
directory, not the root directory. To account for this, I put it in
both.